More python3 conversion, and parted fixes
More byte encoding/decoding.

The print output of parted was changed to not be very useful, but there
is a JSON output that is more useful.

+++ UIpartition/Popup.py
@@ -82,7 +82,7 @@ class Popup:
             
         wlines = nlines - 2
         if (len(lines) < wlines):
-            y += (wlines / 2) - (len(lines) / 2)
+            y += int(wlines / 2) - int(len(lines) / 2)
             wlines = len(lines)
             nlines = wlines
             pass
